% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf_Data.sort.R
\name{abf_Data.sort}
\alias{abf_Data.sort}
\title{Sort field dataset when importing data using importABF}
\usage{
abf_Data.sort(x)
}
\arguments{
\item{x}{}
}
\value{
list of names dfs
}
\description{
Sort field dataset when importing data using importABF
}
\examples{
\dontrun
abf_Data.sort(field

}
