% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf_Data.sort.R
\name{abf_dataSort}
\alias{abf_dataSort}
\title{Sort field dataset when importing data using importABF}
\usage{
abf_dataSort(x)
}
\arguments{
\item{x}{nnest}
}
\value{
list of names dfs
}
\description{
Sort field dataset when importing data using importABF
}
\examples{
\dontrun
abf_Data.sort(field)

}
