% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf_feather.R
\name{abf_feather}
\alias{abf_feather}
\title{All sweeps from sA$ABF to feather}
\usage{
abf_feather(x, units = "ms")
}
\arguments{
\item{x}{}
}
\value{
dfSweeps
}
\description{
All sweeps from sA$ABF to feather
}
\examples{
x = dfs
}
