% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APanalysis.R
\name{APanalysis}
\alias{APanalysis}
\title{Analysis function for APs
Includes a collection of descriptive statistics for each action potential.}
\usage{
APanalysis(
  x,
  rate,
  eONi,
  eOFFi,
  bw = 2,
  minpeakheight = -10,
  minpeakdistance = 0.005 * rate
)
}
\arguments{
\item{x}{sweep vector}

\item{rate}{numeric sampling frequency}

\item{eONi}{numeric index start of epoch check}

\item{minpeakheight}{numeric}

\item{eOFF}{numeric index end of epoch for check}
}
\value{
dataframe
}
\description{
Analysis function for APs
Includes a collection of descriptive statistics for each action potential.
}
\examples{

\dontrun{
x = sweep
x = data[,"data_00075_AD0"]
minpeakheight = -10
rate = 50000
rate = unique(smryDF$sampling_rate)
bw = 2
minpeakheight = -10
minpeakdistance = .005*rate
rate = smryDF$rate[1]
eONi = smryDF$eONi[1]
eOFFi = smryDF$eOFFi[1]
}

}
