% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_Sweeps.R
\name{Avg_Sweeps}
\alias{Avg_Sweeps}
\title{Baseline Analysis for LTD}
\usage{
Avg_Sweeps(x, Wsh = 0, Sw = 20)
}
\arguments{
\item{x}{data.frame}

\item{Wsh}{Wash region to ignore when getting BlAvg}
}
\value{
x
}
\description{
Baseline Analysis for LTD
}
