% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgSweeps.R
\name{avgSweeps}
\alias{avgSweeps}
\title{Baseline Analysis for LTD}
\usage{
avgSweeps(x, obs = "PreC-Bl", nAvg = 20, Wsh = 0)
}
\arguments{
\item{x}{data.frame}

\item{nAvg}{The number of sweeps to observe before}

\item{Wsh}{Wash region to ignore when getting BlAvg}
}
\value{

}
\description{
Baseline Analysis for LTD
}
\examples{
x = field$ABF
}
