% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_adjust.R
\name{baseline_adjust}
\alias{baseline_adjust}
\title{Adjust Baseline}
\usage{
baseline_adjust(x, r)
}
\arguments{
\item{x}{Dataframe to be adjusted}

\item{r}{range of samples to calculate Bl mean from}
}
\description{
Adjust Baseline
}
