% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust.R
\name{bl_adjust}
\alias{bl_adjust}
\title{Baseline adjust by vector. Useful with apply functions. Does not return mV drift}
\usage{
bl_adjust(x, r)
}
\arguments{
\item{x}{vector}

\item{r}{range of samples to calculate Bl mean from}
}
\description{
Baseline adjust by vector. Useful with apply functions. Does not return mV drift
}
