% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bl_Analysis.R
\name{Bl_Analysis}
\alias{Bl_Analysis}
\title{Baseline Analysis for LTD}
\usage{
Bl_Analysis(x, Wsh = 0, out)
}
\arguments{
\item{x}{data.frame}

\item{Wsh}{Wash region to ignore when getting BlAvg}

\item{SA}{postition of the SA}
}
\value{
x
}
\description{
Baseline Analysis for LTD
}
