% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlAnalysis.R
\name{BlAnalysis}
\alias{BlAnalysis}
\title{Baseline Analysis for LTD}
\usage{
BlAnalysis(x, Wsh, slp)
}
\arguments{
\item{x}{data.frame}

\item{Wsh}{Wash region to ignore when getting BlAvg}

\item{slp}{Logical. Apply slope.read function to all sweeps if TRUE.}
}
\value{
x
}
\description{
Baseline Analysis for LTD
}
