% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlAnalysis_0.R
\name{BlAnalysis_0}
\alias{BlAnalysis_0}
\title{Baseline Analysis for LTD}
\usage{
BlAnalysis_0(x, Wsh, SA)
}
\arguments{
\item{x}{data.frame}

\item{Wsh}{Wash region to ignore when getting BlAvg}

\item{SA}{postition of the SA}
}
\value{
x
}
\description{
Baseline Analysis for LTD
}
