% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_proj.R
\name{build_proj}
\alias{build_proj}
\title{Build a project directory}
\usage{
build_proj(x, path = "~")
}
\arguments{
\item{x}{Character Identifiers for project name.}

\item{path}{path for directory}
}
\value{
NA Build project directory
}
\description{
Build a project directory
}
\examples{
\dontrun{
x = "EXA"
build_proj(x)
}
}
