% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccInj.R
\name{ccInj}
\alias{ccInj}
\title{Current clamp inject On
This function identifies where the current is injected}
\usage{
ccInj(x, I = Imon, V = Vmon, testPulse = TRUE, tpPos = "pre", AD = FALSE)
}
\arguments{
\item{x}{a vector. Current signal for current clamp recordings}

\item{testPulse}{Logical. Is there a test pulse in the trace}

\item{tpPos}{Is the test pulse before or after}
}
\value{
list. index points along trace that represents where the current signal turns on and off
}
\description{
Current clamp inject On
This function identifies where the current is injected
}
\examples{
\dontrun{

}
}
