% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccInj_ON.R
\name{ccInj_ON}
\alias{ccInj_ON}
\title{Current clamp inject On
This function identifies where the current clamp is injected}
\usage{
ccInj_ON(x, testPulse = TRUE, tpPos = "pre")
}
\arguments{
\item{x}{a vector. Current signal for current clamp recordings}

\item{testPulse}{Logical. Is there a test pulse in the trace}

\item{tpPos}{}
}
\value{
list. index points along trace that represents where the current signal turns on and off
}
\description{
Current clamp inject On
This function identifies where the current clamp is injected
}
\examples{
\dontrun{

}
}
