% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commentsNWB.R
\name{commentsNWB}
\alias{commentsNWB}
\title{Function intended to separate the comments into a reliable output}
\usage{
commentsNWB(x, sep = "/n")
}
\arguments{
\item{x}{}
}
\description{
Function intended to separate the comments into a reliable output
}
