% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIndex.R
\name{convertIndex}
\alias{convertIndex}
\title{Convert index into timescale by frequency}
\usage{
convertIndex(x, fq)
}
\arguments{
\item{x}{vector of length n}

\item{fq}{sampling frequency in seconds}
}
\description{
Convert index into timescale by frequency
}
