% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateColumns.R
\name{CreateColumns}
\alias{CreateColumns}
\title{Create Columns}
\usage{
CreateColumns(x, n)
}
\arguments{
\item{x}{dataframe}

\item{n}{numer of desired columns}
}
\value{
x
}
\description{
Create Columns
}
