% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/De_Analysis.R
\name{De_Analysis}
\alias{De_Analysis}
\title{Closer analysis and slopes of Decay Data}
\usage{
De_Analysis(x, dfC, SA)
}
\arguments{
\item{x}{Decay data. ncol minimum of 241 including Time}

\item{dfC}{dfCross from BlAnalysis}

\item{SA}{Stim artifact}
}
\value{
x
}
\description{
Closer analysis and slopes of Decay Data
}
