% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeAnalysis.R
\name{DeAnalysis}
\alias{DeAnalysis}
\title{Closer analysis and slopes of Decay Data}
\usage{
DeAnalysis(x, dfC, slp)
}
\arguments{
\item{x}{Decay data. ncol minimum of 241 including Time}

\item{dfC}{dfCross from BlAnalysis}

\item{slp}{Logical. Apply slope.read function to all sweeps if TRUE.}
}
\value{
x
}
\description{
Closer analysis and slopes of Decay Data
}
