% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeAnalysis_0.R
\name{DeAnalysis_0}
\alias{DeAnalysis_0}
\title{Closer analysis and slopes of Decay Data}
\usage{
DeAnalysis_0(x, dfC, SA)
}
\arguments{
\item{x}{Decay data. ncol minimum of 241 including Time}

\item{dfC}{dfCross from BlAnalysis}

\item{SA}{Stim artifact}
}
\value{
x
}
\description{
Closer analysis and slopes of Decay Data
}
