% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_trace.R
\name{df_to_trace}
\alias{df_to_trace}
\title{Plot traces from a dataframe of sweeps.}
\usage{
df_to_trace(
  x,
  sig = "pA",
  msInt = 0.1,
  p0 = 250,
  p1 = 750,
  avg = TRUE,
  multiMax = 1.05,
  multiMin = 1.05,
  scaleBar = 50
)
}
\arguments{
\item{msInt}{the number of samples per ms}
}
\value{

}
\description{
Plot traces from a dataframe of sweeps.
}
