% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMani.R
\name{dfMani}
\alias{dfMani}
\title{dataframe manicure
standardize sweeps to df structure. addfs}
\usage{
dfMani(x, sig = "pA", msInt = 0.1, ROI)
}
\arguments{
\item{sig}{Signal name present in colnames}

\item{msInt}{sample frequency in ms}

\item{ROI}{Region of interest. Need to work on this ways to optimize this.}
}
\value{

}
\description{
dataframe manicure
standardize sweeps to df structure. addfs
}
