% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_ABF.R
\name{dfs_ABF}
\alias{dfs_ABF}
\title{Extract data from ABF as a list of dataframes}
\usage{
dfs_ABF(x)
}
\arguments{
\item{x}{input is sA$ABF}
}
\value{

}
\description{
Extract data from ABF as a list of dataframes
}
\examples{
x = sA$ABF
}
