% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochDF_from_nwb_comments.R
\name{epochDF_from_nwb_comments}
\alias{epochDF_from_nwb_comments}
\title{Return the Epochs in a dataframe format for easier}
\usage{
epochDF_from_nwb_comments(x)
}
\arguments{
\item{x}{Epoch description pulled from extractNWB output comments$Epochs}
}
\value{
dataframe of epochs
}
\description{
Return the Epochs in a dataframe format for easier
}
