% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochReturn.R
\name{epochReturn}
\alias{epochReturn}
\title{Return the Epochs in a dataframe format for easier}
\usage{
epochReturn(x)
}
\arguments{
\item{x}{Epoch description pulled from extractNWB expParas$comments$Epochs}
}
\description{
Return the Epochs in a dataframe format for easier
}
