% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochStimulus.R
\name{epochStimulus}
\alias{epochStimulus}
\title{Get Epoch from stimulus}
\usage{
epochStimulus(x, s, ntail = 2)
}
\arguments{
\item{x}{Epoch description pulled from extractNWB expParas$comments$Epochs}
}
\description{
Get Epoch from stimulus
}
