% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNWB.R
\name{extractNWB}
\alias{extractNWB}
\title{Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{
}}
\usage{
extractNWB(
  x,
  data = FALSE,
  stimulus = FALSE,
  sweeps = NULL,
  acquisition_names = FALSE,
  stimulus_description = FALSE,
  sampling_rate = FALSE,
  comments = FALSE
)
}
\arguments{
\item{x}{path to nwb}

\item{sampling_rate}{}

\item{version}{nwb file version}
}
\description{
Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{
}
}
\examples{
\dontrun{
### V1
x = "exa/QN22.26.029.4A.05.01-compressed.nwb"
version = 1

### V2
x = "exa/QM22.26.031.4A.03.01-compressed.nwb"
version = 2

nwb = rhdf5::H5Fopen(x)

data = TRUE
stimulus = TRUE
sweeps = c(1,2,5,8)
acquisition_names = TRUE
stimulus_description = TRUE
sampling_rate = TRUE
comments = TRUE


}

}
