% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data.R
\docType{data}
\name{field}
\alias{field}
\title{Field Data Set}
\format{
An object of class \code{"cross"}
}
\source{
My own data set outlining a normal field electrophysiology experiment to asses long-term depression in the hippocampus
}
\usage{
data(field)
}
\description{
Field Data Set
}
\examples{
\dontrun{
data(field)
names(field$ABF)
}



}
\keyword{datasets}
