% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileD.R
\name{fileD}
\alias{fileD}
\title{Directory that file sits in}
\usage{
fileD(x, df = FALSE)
}
\arguments{
\item{x}{}
}
\value{
wd
}
\description{
Directory that file sits in
}
\examples{
"~/Projects/yourproj/file.xxx"
}
