% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpks.R
\name{fpks}
\alias{fpks}
\title{Find peaks in a trace}
\usage{
fpks(
  x,
  nups = 1,
  ndowns = nups,
  zero = "0",
  peakpat = NULL,
  minpeakheight = -Inf,
  minpeakdistance = 1,
  threshold = 0,
  npeaks = 0,
  sortstr = FALSE
)
}
\arguments{
\item{sortstr}{}
}
\description{
Find peaks in a trace
}
