% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fthrStim.R
\name{fthrStim}
\alias{fthrStim}
\title{feather a sweep
Feather a sweep by first identifying its stim artifact
then selecting the numer of samples beyond you want.}
\usage{
fthrStim(x, stim, adj = -100, samples = 1500, PP = FALSE)
}
\arguments{
\item{x}{vector}

\item{stim}{Stim artifact. Numeric to identify the point along the 1:length(vector). If blank,}

\item{adj}{Passed to the function stimArtifact() as a UI variable that adjusts the returned value by n samples}

\item{duration}{number of samples to return beyond the stim artifact}
}
\value{
vector
}
\description{
feather a sweep
Feather a sweep by first identifying its stim artifact
then selecting the numer of samples beyond you want.
}
\examples{
\dontrun{

}

}
