% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{generate_project}
\alias{generate_project}
\title{Function used to generate a new project directory.}
\usage{
generate_project(x)
}
\arguments{
\item{x}{Desired name of project}
}
\value{
NA
}
\description{
Function used to generate a new project directory.
}
