% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.pyIPFX.R
\name{get_nwb_version}
\alias{get_nwb_version}
\title{Convert useful ipfx functions to R for portability}
\usage{
get_nwb_version(x)
}
\arguments{
\item{x}{}
}
\description{
Convert useful ipfx functions to R for portability
}
