% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupify.R
\name{groupify}
\alias{groupify}
\title{Group based upon sweep interval.}
\usage{
groupify(x, threshold = 10)
}
\arguments{
\item{threshold}{}
}
\description{
Group based upon sweep interval.
}
\examples{

tst0 =
starting_times = sapply(1:length(tst0), function(x){
tst0[[x]]$starting_time
})

x = starting_times

}
