% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_points.R
\name{identify_points}
\alias{identify_points}
\title{Baseline Analysis for LTD}
\usage{
identify_points(x)
}
\arguments{
\item{x}{data.frame}

\item{Wsh}{Wash region to ignore when getting BlAvg}
}
\value{
x
}
\description{
Baseline Analysis for LTD
}
