% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{Import}
\alias{Import}
\title{Import atf files into the R environment}
\usage{
Import(x, dDir, fmt)
}
\arguments{
\item{x}{The name of a Slice}

\item{fmt}{Format}
}
\value{
dfR
}
\description{
Import atf files into the R environment
}
