% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_ABF.R
\name{Import_ABF}
\alias{Import_ABF}
\title{Import ABF files from experiment.}
\usage{
Import_ABF(x, dir = ".", sv = TRUE)
}
\arguments{
\item{x}{csv metadata file associated with experiment uniquely tagged with Params$mdTag}
}
\value{

}
\description{
Import ABF files from experiment.
}
\examples{
x = "field"
}
