% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importABF.R
\name{importABF}
\alias{importABF}
\title{Import ABF files from experiment.}
\usage{
importABF(
  x,
  mdFlag = "-MD.csv",
  dir = ".",
  format = "abf",
  sv = TRUE,
  ret = TRUE
)
}
\arguments{
\item{x}{Unique ID. Will be combined with mdFlag to isolate and import.}

\item{mdFlag}{tag that will identify md from experimental parameters}

\item{dir}{base directory to start search in}

\item{format}{what format of file are you importing.}

\item{sv}{Logical. Save nnest. Default is TRUE. Will return list of elements. Named as nnest.}

\item{ret}{= return the list or just save it.}
}
\value{
list with all the components to run data analysis
}
\description{
Import ABF files from experiment.
}
\examples{
\dontrun{
field = importABF("field", mdFlage = "-MD.csv", dir = "exa/field")
}
}
