% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isolate.R
\name{Isolate}
\alias{Isolate}
\title{Isolate Bl sweeps from wash in sweeps if relevant}
\usage{
Isolate(x, BlR, Wsh)
}
\arguments{
\item{x}{data}

\item{BlR}{Baseline - defaults to 80}

\item{Wsh}{Wash region - defaults to 0}
}
\value{
Sweeps
}
\description{
Isolate Bl sweeps from wash in sweeps if relevant
}
