% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateAPs.R
\name{isolateAPs}
\alias{isolateAPs}
\title{Title}
\usage{
isolateAPs(x, fp, rate, eONi, eOFFi, seg_term = 0.05)
}
\arguments{
\item{x}{sweep vector}
}
\description{
Title
}
\examples{

hct = loadHCT("Q20.26.005.1A.02.02")
sweep = extractNWB(hct$files$nwb, sweeps = hct$ffts$data_00036_AD0$trace, slim = TRUE)
rate = hct$ffts$data_00036_AD0$rate
plot(sweep, type = "l")

x = sweep
seg_term = 0.05

}
