% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolateSegment.R
\name{isolateSegment}
\alias{isolateSegment}
\title{Isolate action potential and subsequent interstimulus segment
removes NA values at the end that are residual from data frame
selects the region after the AHP}
\usage{
isolateSegment(x, p0, pN)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
numeric vector with na removed and segmented from p0 to end of trace.
}
\description{
Isolate action potential and subsequent interstimulus segment
removes NA values at the end that are residual from data frame
selects the region after the AHP
}
