% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmROI.R
\name{lmROI}
\alias{lmROI}
\title{Title}
\usage{
lmROI(x, p1, p2)
}
\arguments{
\item{x}{region of interest}

\item{p1}{first point to investigate}

\item{p2}{second point to investigate}
}
\value{
x
}
\description{
Title
}
