% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rda.R
\name{load_rda}
\alias{load_rda}
\title{Load rda nest into Global Env}
\usage{
load_rda(x, dir = ".", ID = "-nphys", select = "FALSE")
}
\arguments{
\item{x}{Unique identifier for data nest}

\item{dir}{top level directory to start at for searching for data nest.}

\item{ID}{rda identifier. Used to distinguish between experiment types and is based
upon the name of the rda file. must be flagged.}

\item{select}{}
}
\value{
Metadata
}
\description{
Load rda nest into Global Env
}
