% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTDpchange.R
\name{LTDpchange}
\alias{LTDpchange}
\title{Plot of LTD over time}
\usage{
LTDpchange(x, b, d, ret = TRUE, ggPrint = TRUE)
}
\arguments{
\item{x}{string of values}

\item{b}{baseline sweeps to analyze against}
}
\description{
Plot of LTD over time
}
