% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_proj.R
\name{new_proj}
\alias{new_proj}
\title{Function used to generate a new project directory.}
\usage{
new_proj(x)
}
\arguments{
\item{x}{Desired name of project}
}
\value{
NA
}
\description{
Function used to generate a new project directory.
}
