% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newExp.R
\name{newExp}
\alias{newExp}
\title{newExp
Add a new experiment to the project
Draws on the exp/params.ods file to append an experiment}
\usage{
newExp(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
newExp
Add a new experiment to the project
Draws on the exp/params.ods file to append an experiment
}
