% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDGDev-package.R
\docType{package}
\name{nphys-package}
\alias{nphys}
\alias{nphys-package}
\title{nphys: R package for interpreting neurphysiology field and whole-cell data}
\description{
Intended to be used for reproducabile analysis and presentation of field and whole cell data collected in neurophysiology experiments
}
\author{
\strong{Maintainer}: Scott Sawchuk \email{ssawchuk@uvic.ca}

}
\keyword{internal}
