% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwbPlot_sweep.R
\name{nwbPlot_sweep}
\alias{nwbPlot_sweep}
\title{Title}
\usage{
nwbPlot_sweep(
  x,
  acquisition_name,
  stimulus_name = FALSE,
  ds_int = NULL,
  return_plot = FALSE,
  scaleBar = TRUE,
  show_plot = TRUE,
  sweepColour = "black",
  return_df = FALSE
)
}
\arguments{
\item{x}{}
}
\description{
Title
}
\examples{
x = "/exd/nwb/NWBv2.nwb"



}
