% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_BlvDe.R
\name{plot.BlvsDecay}
\alias{plot.BlvsDecay}
\title{Print baseline averaged trace}
\usage{
\method{plot}{BlvsDecay}(x)
}
\arguments{
\item{x}{}
}
\value{
gg
}
\description{
Print baseline averaged trace
}
\examples{
x = field$ABF
}
