% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_BlvDe.R
\name{plot_BlvsDe}
\alias{plot_BlvsDe}
\title{Print baseline averaged trace}
\usage{
plot_BlvsDe(x)
}
\arguments{
\item{Fq}{}
}
\value{
gg
}
\description{
Print baseline averaged trace
}
\examples{
x = field$traces
}
