% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBlvDecay.R
\name{plotBlvDecay}
\alias{plotBlvDecay}
\title{Basic Baseline vs Decay plot.}
\usage{
plotBlvDecay(x, Wsh = FALSE)
}
\arguments{
\item{x}{Basic project nested list that nphys functions can run off}
}
\value{
gg
}
\description{
Basic Baseline vs Decay plot.
}
\examples{
\dontrun{
x = field$ABF
plotBlvDecay(x)
}
}
