% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoid.R
\name{plotVoid}
\alias{plotVoid}
\title{Plot void trace}
\usage{
plotVoid(x, plot = TRUE, return = TRUE)
}
\arguments{
\item{x}{numerical: a vector}

\item{plot}{logical: show plot if TRUE}

\item{return}{logical: returns object if TRUE}
}
\description{
Plot void trace
}
