% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVoidNWB.R
\name{plotVoidNWB}
\alias{plotVoidNWB}
\title{Plot void trace pulled}
\usage{
plotVoidNWB(x, sweeps, plot_sweep = TRUE, return_sweep = FALSE)
}
\arguments{
\item{x}{file path}

\item{sweeps}{numeric or character string}

\item{plot_sweep}{do you want the sweep plotted?}

\item{return_sweep}{do you want to return the df used to create the sweep}
}
\description{
Plot void trace pulled
}
\examples{
x = "~/nphys/exd/nwb/NWBv2.nwb"
sweeps = 6

}
