% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pZero.R
\name{pZero}
\alias{pZero}
\title{pZero}
\usage{
pZero(x, n = 1)
}
\arguments{
\item{x}{numeric vector along which you can select a number of point}

\item{n}{how many points do you want to choose?}

\item{sel}{}
}
\description{
pZero
}
\examples{
x =

}
