% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticABF.R
\name{reticABF}
\alias{reticABF}
\title{Use reticulate and python to extract abf data}
\usage{
reticABF(x)
}
\arguments{
\item{x}{}
}
\description{
Use reticulate and python to extract abf data
}
\examples{
x = loadD2AT(Rds = FALSE)
x = nnest

}
