% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sA_LFS.R, R/sAnalysis_2060.R
\name{sA_LFS}
\alias{sA_LFS}
\title{Analysis by cursor placement}
\usage{
sA_LFS(x, m = 1)

sA_LFS(x, m = 1)
}
\arguments{
\item{x}{}

\item{x}{}
}
\value{
dtSw

dtSw
}
\description{
Analysis by cursor placement

Analysis with cursors at 10% and 50% of the observable EPSP
}
