% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAnalysis_1050.R
\name{sAnalysis_1050}
\alias{sAnalysis_1050}
\title{Analysis with cursors at 10% and 50% of the observable EPSP}
\usage{
sAnalysis_1050(x)
}
\arguments{
\item{x}{}
}
\value{
ndt
}
\description{
Analysis with cursors at 10% and 50% of the observable EPSP
}
