% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAnalysis_2060.R
\name{sAnalysis_2060Sw}
\alias{sAnalysis_2060Sw}
\title{Analysis with cursors at 10% and 50% of the observable EPSP}
\usage{
sAnalysis_2060Sw(x)
}
\arguments{
\item{x}{}
}
\value{
dtSw
}
\description{
Analysis with cursors at 10% and 50% of the observable EPSP
}
