% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAnalysis_LFS.R
\name{sAnalysis_LFS}
\alias{sAnalysis_LFS}
\title{Analysis with cursors at 10% and 50% of the observable EPSP}
\usage{
sAnalysis_LFS(x)
}
\arguments{
\item{x}{}
}
\value{
dtSw
}
\description{
Analysis with cursors at 10% and 50% of the observable EPSP
}
