% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAnalysis_maxSlope.R
\name{sAnalysis_maxSlope}
\alias{sAnalysis_maxSlope}
\title{Analysis with cursors at 10% and 50% of the observable EPSP}
\usage{
sAnalysis_maxSlope(x)
}
\arguments{
\item{x}{}
}
\value{
ndt_2
}
\description{
Analysis with cursors at 10% and 50% of the observable EPSP
}
