% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot_LFS.R
\name{ScatterPlot_LFS}
\alias{ScatterPlot_LFS}
\title{Create Scatterplot of LTD traces}
\usage{
ScatterPlot_LFS(x, j)
}
\arguments{
\item{x}{meltedDF}

\item{j}{Pooled Treatment AgeGroup to investigate}
}
\value{
NA
}
\description{
Create Scatterplot of LTD traces
}
