% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot_Tr.R
\name{ScatterPlot_Tr}
\alias{ScatterPlot_Tr}
\title{Scatter plots for treatment effects at AgeGroups}
\usage{
ScatterPlot_Tr(x, j, s)
}
\arguments{
\item{x}{melted df with variable and value as x and y components}

\item{j}{AgeGroup}

\item{s}{Scatter shape of treatment}
}
\value{
ggplot
}
\description{
Scatter plots for treatment effects at AgeGroups
}
