% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot_Tr_LFS.R
\name{ScatterPlot_Tr_LFS}
\alias{ScatterPlot_Tr_LFS}
\title{Scatter plots for treatment effects at AgeGroups}
\usage{
ScatterPlot_Tr_LFS(x, j, s, lb)
}
\arguments{
\item{x}{melted df with variable and value as x and y components}

\item{j}{AgeGroup}

\item{s}{Scatter shape of treatment}

\item{lb}{text}
}
\value{
ggplot
}
\description{
Scatter plots for treatment effects at AgeGroups
}
