% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentFFT.R
\name{segmentFFT}
\alias{segmentFFT}
\title{Function for applying a FFT to a dataframe of segments}
\usage{
segmentFFT(x, p0, pN, rate, lower = 1, upper = 1000, melt = TRUE)
}
\arguments{
\item{x}{hct$ffts$\link{trace}$APsweeps or hct$ffts$\link{trace}$APsweeps\link{,APn}}

\item{p0}{head of trace point to remove from fft analysis}

\item{lower}{lower bound of fft analysis}

\item{upper}{upper bound of fft analysis}

\item{melt}{determines whether melted dataframe or datavector is returned}
}
\value{
returns the fqMelt or datavector depending on melt variable in function
}
\description{
Function for applying a FFT to a dataframe of segments
}
\examples{

x = seg01

}
