% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentFilter.R
\name{segmentFilter}
\alias{segmentFilter}
\title{Apply a filter to a segment length.}
\usage{
segmentFilter(x, rate, lower = 1, upper = 2000, plot = FALSE, return = FALSE)
}
\arguments{
\item{x}{segment to be filtered}

\item{rate}{rate of acquisition}

\item{lower}{lower bound}

\item{upper}{uppper bound}

\item{plot}{Do you want to plot the figure.}
}
\value{
plot or vector
}
\description{
Apply a filter to a segment length.
}
\examples{

\dontrun{
x = seg01
}

}
