% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAnalysis_2060.R
\name{slope_Analysis}
\alias{slope_Analysis}
\title{Analysis with cursors at 10% and 50% of the observable EPSP}
\usage{
slope_Analysis(x, m = 1)
}
\arguments{
\item{x}{}
}
\value{
dtSw
}
\description{
Analysis with cursors at 10% and 50% of the observable EPSP
}
