% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stim.artifcat.R
\name{stim.Artifact}
\alias{stim.Artifact}
\title{Find stim artifact for traces with single}
\usage{
stim.Artifact(x)
}
\arguments{
\item{x}{numerical vector to be examined}
}
\value{
fp
}
\description{
Find stim artifact for traces with single
}
