% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimArtifcat.R
\name{stimArtifact}
\alias{stimArtifact}
\title{Utilize the point along a trace
where the rate of change between points
is the greatest to identify the stim artifact.}
\usage{
stimArtifact(x, adj = 0)
}
\arguments{
\item{x}{Vector along with which to examine. Currently only supports one stim artifact per trace}

\item{adj}{Make adjustments to how far up or down the trace from the stim artifact you want returned}
}
\value{
index along trace that matches the point where the difference between points is the greatest to identify the on phase of the stim artifact.
This is necessary when trying to locate the response along the trace, but not having the Stim channel data.
}
\description{
Utilize the point along a trace
where the rate of change between points
is the greatest to identify the stim artifact.
}
\examples{
\dontrun{
stimArtifact(field$traces$Bl_avg)
}
}
