% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepRead.R
\name{sweepRead}
\alias{sweepRead}
\title{Adjust Baseline}
\usage{
sweepRead(x, r)
}
\arguments{
\item{x}{Dataframe to be adjusted}

\item{r}{range of samples to calculate Bl mean from}
}
\description{
Adjust Baseline
}
