% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPulse.R
\name{testPulse}
\alias{testPulse}
\title{Test pulse analysis}
\usage{
testPulse(v, i, dV = NA, dI = NA)
}
\arguments{
\item{v}{voltage trace}

\item{i}{current trace}

\item{dV}{magnitude of test pulse in Volts.}

\item{dI}{magnitude of test pulse in Amps.}
}
\description{
Test pulse analysis
}
