% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_downsampling.R
\name{threshold_downsampling}
\alias{threshold_downsampling}
\title{Downsample based on rate of change between points. Can greatly reduce the number of points in a trace}
\usage{
threshold_downsampling(signal, threshold)
}
\arguments{
\item{signal}{sweep vector}

\item{threshold}{absolute value of difference between signal points.}
}
\description{
Downsample based on rate of change between points. Can greatly reduce the number of points in a trace
}
