% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmAvg.R
\name{vmAvg}
\alias{vmAvg}
\title{Get voltage average from 0 for sweeps}
\usage{
vmAvg(x, r = 1:1000)
}
\arguments{
\item{x}{numeric vector}

\item{r}{range of samples to calculate mean from}
}
\description{
Get voltage average from 0 for sweeps
}
