% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_exp_metadataNWB.R
\name{write_exp_metadataNWB}
\alias{write_exp_metadataNWB}
\title{Write experimental metadata critical for dandi upload to an nwb file}
\usage{
write_exp_metadataNWB(nwb_file, subject_id, age, sex, species)
}
\arguments{
\item{nwb_file}{nwb file path}

\item{subject_id}{character}

\item{age}{character ISO 8601 format}

\item{sex}{character}

\item{species}{character latin specific species designation}
}
\value{
NA
}
\description{
Write experimental metadata critical for dandi upload to an nwb file
}
\examples{
\dontrun{
nwb_file = "~/Q21.26.005.11.03.04-compressed.nwb"
subject_id = "Q21.26.005"
age = "M"
sex = "P5Y90D"
species = "macaca nemestrina"



}








}
