% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrkD.R
\name{wrkD}
\alias{wrkD}
\title{Directory that file sits in}
\usage{
wrkD(x, df = FALSE)
}
\arguments{
\item{x}{}
}
\value{
wd
}
\description{
Directory that file sits in
}
\examples{
"~/Projects/yourproj/file.xxx"
}
