% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroAdjust.R
\name{zeroAdjust}
\alias{zeroAdjust}
\title{Zero adjust. Useful with apply functions.}
\usage{
zeroAdjust(x, r = 1:1000)
}
\arguments{
\item{x}{numeric vector}

\item{r}{range of samples to calculate mean from}
}
\description{
Zero adjust. Useful with apply functions.
}
