% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abf_Data.sort.R
\name{abf_Data.sort}
\alias{abf_Data.sort}
\title{Sort js dataset from sA after import using readABF.}
\usage{
abf_Data.sort(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Sort js dataset from sA after import using readABF.
}
