% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APanalysis.R
\name{APanalysis}
\alias{APanalysis}
\title{Analysis function for APs
Includes a collection of descriptive statistics for each action potential.}
\usage{
APanalysis(
  x,
  rate,
  eONi,
  eOFFi,
  bw = 2,
  minpeakheight = -10,
  minpeakdistance = 0.005 * rate
)
}
\arguments{
\item{x}{sweep vector}

\item{eOFF}{}
}
\description{
Analysis function for APs
Includes a collection of descriptive statistics for each action potential.
}
\examples{
x = sweep
x = data[,"data_00075_AD0"]
minpeakheight = -10
rate = 50000
rate = rate
bw = 4
minpeakdistance = .01*rate

}
