% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APanalysis.R
\name{APanalysis}
\alias{APanalysis}
\title{Title}
\usage{
APanalysis(
  x,
  rate,
  eONi,
  eOFFi,
  bw = 2,
  minpeakheight = -10,
  minpeakdistance = 0.005 * rate
)
}
\arguments{
\item{x}{sweep vector}

\item{eOFF}{}
}
\description{
Title
}
\examples{
x = nphys
x = ksweep
minpeakheight = -10
rate = 50000
rate = rate
bw = 4
minpeakdistance = .01*rate

}
