% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust.R
\name{bl_adjust}
\alias{bl_adjust}
\title{Adjust Baseline}
\usage{
bl_adjust(x, r)
}
\arguments{
\item{x}{Dataframe to be adjusted}

\item{r}{range of samples to calculate Bl mean from}
}
\description{
Adjust Baseline
}
