% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust.R
\name{bl_adjust}
\alias{bl_adjust}
\title{Get voltage drift from 0 for sweeps}
\usage{
bl_adjust(x, r)
}
\arguments{
\item{x}{Dataframe of sweeps}

\item{r}{range of samples to calculate Bl mean from}
}
\description{
Get voltage drift from 0 for sweeps
}
