% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIndex.R
\name{convertIndex}
\alias{convertIndex}
\title{Convert index into Seconds by frequency}
\usage{
convertIndex(x, fq)
}
\arguments{
\item{x}{vector of length n}

\item{fq}{sampling frequency in seconds}
}
\value{
vector
}
\description{
Convert index into Seconds by frequency
}
\examples{

\dontrun{

}


}
