% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_ABF.R
\name{dfs_ABF}
\alias{dfs_ABF}
\title{Extract data or other variables from the imported ABF data.}
\usage{
dfs_ABF(x, int = "data", select = FALSE, returnList = TRUE)
}
\arguments{
\item{x}{data from import}

\item{int}{Character
Indicate what your interested in extracting from the contents of the imported ABF format.}

\item{select}{Logical or Numeric
Indicate if you want to choose from a list of options to be returned.
Options to choose from are contents of ABF file after import.}

\item{returnList}{Logical.
If TRUE, will return elements as a nested list.
Useful when extracting the nested dataframes that make up the majority of the nested data.
If FALSE, will unlist the elements ard return data as a single vector.
Useful if wanting to return something like sampling rate or signal channels.}
}
\value{
Extracted elements as determined by the user input at int. Default is "data".
}
\description{
Extract data or other variables from the imported ABF data.
}
\examples{
\dontrun{
x = field$ABF
dfs <- dfs_ABF(x)
}
}
