% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs_ABF.R
\name{dfs_ABF}
\alias{dfs_ABF}
\title{Extract data or other variables from imported ABF format}
\usage{
dfs_ABF(x, int = "data", select = FALSE, returnList = TRUE)
}
\arguments{
\item{x}{data from import}

\item{int}{character
Set the option for extract}

\item{sel}{logical
Determine if you want to choose from a list of options to be returned. Ca}
}
\value{
list of extracted elements
}
\description{
Extract data or other variables from imported ABF format
}
\examples{
x = field$ABF
dfs <- dfs_ABF(x)
}
