% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvdt_mV.R
\name{dvdt_mV}
\alias{dvdt_mV}
\title{Display figure of AP dV/dt}
\usage{
dvdt_mV(x, rate, plot = TRUE)
}
\arguments{
\item{x}{AP vector}

\item{rate}{sampling frequency}
}
\description{
Display figure of AP dV/dt
}
\examples{

 #AP = x
#sweep = hct$ffts$data_00029_AD0$sweep
# fp = hct$ffts$data_00029_AD0$peaks
# AP = sweep[fp[4,3],fp[4,4]]



}
