% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochReturn.R
\name{epochReturn}
\alias{epochReturn}
\title{Return the Epochs in a dataframe format for easier building
Built in projHCT ported over 240508}
\usage{
epochReturn(x)
}
\arguments{
\item{x}{Epoch description pulled from extractNWB exp$comments$Epochs}
}
\description{
Return the Epochs in a dataframe format for easier building
Built in projHCT ported over 240508
}
