% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epochStimulus.R
\name{epochStimulus}
\alias{epochStimulus}
\title{Return the Epochs in a dataframe format for easier}
\usage{
epochStimulus(x, s, ntail = 2)
}
\arguments{
\item{x}{Epoch description pulled from extractNWB expParas$comments$Epochs}
}
\description{
Return the Epochs in a dataframe format for easier
}
