% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNWB.R
\name{extractNWB}
\alias{extractNWB}
\title{Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{

Utilizes the rhdf5 package to load nwb object into environment and follows
sequence of directive to extract data.
}}
\usage{
extractNWB(
  x,
  acquisition = FALSE,
  sweeps = NULL,
  stimulus_sweeps = NULL,
  epochs = TRUE,
  slim = FALSE
)
}
\arguments{
\item{x}{path to nwb}

\item{data}{return timeseries in NWB}

\item{sampling_rate}{}
}
\description{
Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{

Utilizes the rhdf5 package to load nwb object into environment and follows
sequence of directive to extract data.
}
}
\examples{
\dontrun{
### V1
x = "data-raw/NWBv1/NWBv1.nwb"
version = 1

### V2
x = "data-raw/NWBv2/NWBv2.nwb"
version = 2

nwb = rhdf5::H5Fopen(x)

acquisition = FALSE
stimulus = TRUE
sweeps = c(1,2,5,8)
acquisition_names = TRUE
stimulus_description = TRUE
sampling_rate = TRUE
sweeps = NULL
comments = TRUE
epochs = TRUE


}

}
