% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNWB.R
\name{extractNWB}
\alias{extractNWB}
\title{Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{
}}
\usage{
extractNWB(
  x,
  version = NULL,
  data = FALSE,
  stimulus = FALSE,
  sweeps = NULL,
  sweep_names = FALSE,
  stimulus_description = FALSE,
  sampling_rate = FALSE
)
}
\arguments{
\item{x}{path to nwb}

\item{version}{nwb file version}

\item{sampling_rate}{}
}
\description{
Extract data from nwb file
\subsection{}{

This function is utilized by the stimSets.R data wrangler script to extract the names of
protocols used across the datasets being investigated.
}

\subsection{}{
}
}
\examples{
\dontrun{
x = "exa/QM22.26.031.4A.03.01-compressed.nwb"
nwb = rhdf5::H5Fopen(x)
}

}
