% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-data.R
\docType{data}
\name{field}
\alias{field}
\title{Field Data Set}
\format{
a \code{nested.list} with 5 elements:
\describe{
\item{wd}{working directory}
\item{md}{Experimental metadata}
\item{files}{Imported files}
\item{ABF}{Imported data from files}
\item{traces}{example traces}
}
}
\source{
Christie Laboratory, Victoria B.C.
}
\usage{
data(field)
}
\description{
My own data set outlining a normal field electrophysiology experiment to asses long-term depression in the hippocampus
}
\keyword{datasets}
