% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_points.R
\name{identify_points}
\alias{identify_points}
\title{Click to identify points along a trace.}
\usage{
identify_points(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
x
}
\description{
Click to identify points along a trace.
}
