% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{Import}
\alias{Import}
\title{Import atf files into the R environment}
\usage{
Import(x, dDir = ".", fmt = "atf")
}
\arguments{
\item{x}{The name of a Slice}

\item{dDir}{directory to search from}

\item{fmt}{File format to import - must be in its own folder named for the format.}
}
\value{
dfR
}
\description{
Import atf files into the R environment
}
