% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABF_Import.R
\name{Import_ABF}
\alias{Import_ABF}
\title{Import ABF files from experiment.}
\usage{
Import_ABF(Slice, lfdir = ".", sv = TRUE)
}
\arguments{
\item{Slice}{metadata associated with unique slice tagged with -SliceMD.csv}
}
\value{

}
\description{
Import ABF files from experiment.
}
\examples{
x = "exa/field"
}
