% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsolateWsh.R
\name{IsolateWsh}
\alias{IsolateWsh}
\title{Isolate relevant columns from wash in sweeps if relevant}
\usage{
IsolateWsh(x, Wsh = 0, BlR = 80)
}
\arguments{
\item{x}{data in the form of}

\item{Wsh}{Wash region - defaults to 0}

\item{BlR}{Baseline - defaults to 80}
}
\value{

}
\description{
Isolate relevant columns from wash in sweeps if relevant
}
\examples{
x =
}
