% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwbPlot_sweep.R
\name{nwbPlot_sweep}
\alias{nwbPlot_sweep}
\title{Plot a sweep from the nwb file path}
\usage{
nwbPlot_sweep(
  x,
  acquisition_name,
  stimulus_name = FALSE,
  ds_int = NULL,
  return_plot = FALSE,
  scaleBar = TRUE,
  show_plot = TRUE,
  sweepColour = "black",
  return_df = FALSE
)
}
\arguments{
\item{x}{nwb file path}

\item{acquisition_name}{character string or numeric}

\item{stimulus_name}{character}

\item{return_df}{}
}
\description{
Plot a sweep from the nwb file path
}
\examples{
x = "/home/nrsc/nphys/exd/nwb/NWBv2.nwb"



}
