% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSweeps.R
\name{pullSweeps}
\alias{pullSweeps}
\title{Pull specific sweeps from imported data}
\usage{
pullSweeps(
  x,
  pull = "PreC-Bl",
  select = FALSE,
  zero = TRUE,
  feather = TRUE,
  adjNeg = -100,
  adjPos = 1400
)
}
\arguments{
\item{x}{field$ABF}

\item{pull}{Part of the experiment you want to pull as determined by names(field$ABF)}

\item{select}{logical
If TRUE, will prompt with options to select.}

\item{zero}{adjust baseline of traces to zero. Is ignored if feather is set to FALSE}

\item{feather}{logical to determine if you want to return remove the points in front of the stim artifact. If FALSE, a list will be returned}

\item{adjNeg}{how much to add to the stim artifact for the start point when observing the sweep. Default is -100 samples}

\item{adjPos}{how much to add to the stim artifact for the end point when observing the sweep. Default is 1400 samples}
}
\value{
dataframe of sweeps unless feather is FALSE
}
\description{
Pull specific sweeps from imported data
}
\examples{
\dontrun{
x = field$ABF \%>\% pullSweeps()
}
}
