% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullSweeps.R
\name{pullSweeps}
\alias{pullSweeps}
\title{Pull specific sweeps from imported data}
\usage{
pullSweeps(x, pull = "PreC-Bl", select = FALSE, zero = TRUE)
}
\arguments{
\item{pull}{Sweeps from the part of the experiment you want to pull}

\item{select}{logical
If TRUE, will prompt with options to select.}
}
\value{

}
\description{
Pull specific sweeps from imported data
}
\examples{
x = field$ABF
}
