% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimArtifcat.R
\name{stimArtifact}
\alias{stimArtifact}
\title{Identify the point along a trace where the rate of change between points is the greatest to identify the stim artifact.}
\usage{
stimArtifact(x, pre = 100)
}
\arguments{
\item{x}{Vector along with which to examine. Currently only supports one SA per trace}
}
\value{

}
\description{
Identify the point along a trace where the rate of change between points is the greatest to identify the stim artifact.
}
