% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimArtifcat.R
\name{stimArtifact}
\alias{stimArtifact}
\title{Utilize the point along a trace
where the rate of change between points
is the greatest to identify the stim artifact.}
\usage{
stimArtifact(x, adj = 0, headORtail = "head", n = 1)
}
\arguments{
\item{x}{Vector along with which to examine. Currently only supports one SA per trace}
}
\value{

}
\description{
Utilize the point along a trace
where the rate of change between points
is the greatest to identify the stim artifact.
}
\examples{
x = field$traces$Bl_avg
}
