% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrkD.R
\name{wrkD}
\alias{wrkD}
\title{Return directory that file is in}
\usage{
wrkD(x)
}
\arguments{
\item{x}{a path}
}
\value{
wd
}
\description{
Return directory that file is in
}
\examples{
\dontrun{
wrkD("~/proj/yourproj/fi.le")
}
}
