% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrkD.R
\name{wrkD}
\alias{wrkD}
\title{Return file path without last path element.}
\usage{
wrkD(x)
}
\arguments{
\item{x}{a path}
}
\value{
wd
}
\description{
Return file path without last path element.
}
\examples{
\dontrun{
wrkD("~/proj/yourproj/fi.le")
}
}
