% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroAdjust.R
\name{zeroAdjust}
\alias{zeroAdjust}
\title{Zero adjust.}
\usage{
zeroAdjust(x, r = 1:1000)
}
\arguments{
\item{x}{numeric vector}

\item{r}{range of samples to calculate mean from}
}
\description{
Used with apply functions accross columns.
Calculation done on each sweep as a vector.
}
\examples{
\dontrun{
Using the example dataset 'field'
dfs_ABF(field$ABF)[[1]][[1]] \%>\% zeroAdjust()

}
}
